add enemy anim, game reset when collide with player 	renamed:    Assets/Scripts.meta -> Assets/Enemy.meta 	new file:   Assets/Enemy/1.PNG 	new file:   Assets/Enemy/1.PNG.meta 	new file:   Assets/Enemy/1_0.controller 	new file:   Assets/Enemy/1_0.controller.meta 	new file:   Assets/Enemy/2.PNG 	new file:   Assets/Enemy/2.PNG.meta 	new file:   Assets/Enemy/3.PNG 	new file:   Assets/Enemy/3.PNG.meta 	new file:   Assets/Enemy/4.PNG 	new file:   Assets/Enemy/4.PNG.meta 	new file:   Assets/Enemy/5.PNG 	new file:   Assets/Enemy/5.PNG.meta 	new file:   Assets/Enemy/6.PNG 	new file:   Assets/Enemy/6.PNG.meta 	new file:   Assets/Enemy/7.PNG 	new file:   Assets/Enemy/7.PNG.meta 	new file:   Assets/Enemy/8.PNG 	new file:   Assets/Enemy/8.PNG.meta 	new file:   Assets/Enemy/Enemy.anim 	new file:   Assets/Enemy/Enemy.anim.meta 	new file:   Assets/Enemy/EnemyChase.cs 	renamed:    Assets/Scripts/EnemyChase.cs.meta -> Assets/Enemy/EnemyChase.cs.meta 	deleted:    Assets/Scripts/EnemyChase.cs
