+++ Assets/Enemy/1_0.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b9c99783c3e124d41b811837c8f03936
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
