+++ Assets/Enemy/Enemy.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 45ea8d3a21e1f1541a2736ab053d2249
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
